Refactor pod name to separate file
