+++ workflow/util/util.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"hash/fnv"
 	"io/ioutil"
 	"math/rand"
 	"net/http"
@@ -51,11 +50,6 @@ import (
 	"github.com/argoproj/argo-workflows/v3/workflow/packer"
 	"github.com/argoproj/argo-workflows/v3/workflow/templateresolution"
 	"github.com/argoproj/argo-workflows/v3/workflow/validate"
-)
-
-const (
-	maxK8sResourceNameLength = 253
-	k8sNamingHashLength      = 10
 )
 
 // NewWorkflowInformer returns the workflow informer used by the controller. This is actually
@@ -1060,32 +1054,4 @@ func GetNodeType(tmpl *wfv1.Template) wfv1.NodeType {
 		return wfv1.NodeTypeSuspend
 	}
 	return ""
-}
-
-// PodName return a deterministic pod name
-func PodName(workflowName, nodeName, templateName, nodeID string) string {
-	if os.Getenv("USE_LEGACY_POD_NAMES") == "true" {
-		return nodeID
-	}
-
-	if workflowName == nodeName {
-		return workflowName
-	}
-
-	prefix := fmt.Sprintf("%s-%s", workflowName, templateName)
-	prefix = ensurePodNamePrefixLength(prefix)
-
-	h := fnv.New32a()
-	_, _ = h.Write([]byte(nodeName))
-	return fmt.Sprintf("%s-%v", prefix, h.Sum32())
-}
-
-func ensurePodNamePrefixLength(prefix string) string {
-	maxPrefixLength := maxK8sResourceNameLength - k8sNamingHashLength
-
-	if len(prefix) > maxPrefixLength-1 {
-		return prefix[0 : maxPrefixLength-1]
-	}
-
-	return prefix
 }
